+++ repro/models/deutsch2021.py
@@ -1,0 +1,8 @@
+from ._models.deutsch2021.src.dataset_reader import (
+    Deutsch2021QuestionAnsweringEvaluationDatasetReader,
+)
+from ._models.deutsch2021.src.models import (
+    QAEvalQuestionAnsweringModel,
+    QAEvalQuestionGenerationModel,
+)
+from ._models.deutsch2021.src.setup import Deutsch2021SetupSubcommand