+++ repro/models/sacrerouge.py
@@ -1,0 +1,2 @@
+from ._models.sacrerouge.src.models import ROUGE
+from ._models.sacrerouge.src.setup import SacreROUGESetupSubcommand