+++ tests/data/output_writers/default_test.py
@@ -23,3 +23,25 @@ class TestDefaultOutputWriter(unittest.TestCase):
                 {"instance_id": "1", "model_id": "test-model", "prediction": "a"},
                 {"instance_id": "2", "model_id": "test-model", "prediction": "b"},
             ]
+
+    def test_unequal_number_of_instances_predictions(self):
+        writer = DefaultOutputWriter()
+        with TemporaryDirectory() as temp:
+            output_file = f"{temp}/output.jsonl"
+
+            instances = [
+                {"instance_id": "1"},
+                {"instance_id": "2"},
+            ]
+            predictions = ["a", "b", "c"]
+            with self.assertRaises(Exception):
+                writer.write(instances, predictions, output_file, "test-model")
+
+            instances = [
+                {"instance_id": "1"},
+                {"instance_id": "2"},
+                {"instance_id": "3"},
+            ]
+            predictions = ["a", "b"]
+            with self.assertRaises(Exception):
+                writer.write(instances, predictions, output_file, "test-model")